sw/requirements.txt
migen>=1.0
pyusb>=1.0.0b2
